Implemented the layout of the drawer for adding expenses
In this commit, i worked on the ui layout of the drawer responsible for adding new expense entries, the ui is based on shadcn/ui components. Also created two mockup cards for the expenses page, an animated title header and top bar, with a settings dropmenu for adjusting the layout of the expenses page (logic implementation for the settings is not yet done).