Expenses Page Layout and Add Expense Drawer backend and frontend logic implementation
in this commit, i completed the layout of the expenses page, with three cards so far, and with filters for the details card functional, as well as the ability to customise the layout by toggling on and off each card. I also implmented the logic of filling, submitting and storing new expense entries. TODO : - Once a new expense is successfully stored, it triggers a full page reload, which is not great for keeping the state of the drawer and react intact, solution would be maybe to add a context provider wrapper that handles this state and that wouldn't be impacted by Inertia triggering react to reload. TODO : handle errors on the front end if the expense submission errors, by showing the origin of the error, and adequat ui feedback. TODO : maybe implement a toast for showing a successful expense submission.