Created the setup pages : spenders,categories, places and ui interactivity for selection.
Not yet done : sending data to the backend and redirecting as needed. Next : create the setup_progress table for managing setup progress of users (to be used to redirect the user proprely). then implement the table in the logic of the app. then implementthe logic for adding spenders, categories, and places from the setup pages (the SpenderController, CategoryController and PlaceController are already created, but no logic implemnted yet)