Implemented the logic for account setup after registring. And refactored the expenses page to use the new database architecture
In this commit, i worked on the three setup steps, implemnted the logic for moving from one step to another, using the setup progress table, updated the middleware to take care of redirecting to the proper route. Reworked the expense page to use the new data structure and database architecture for getting and submitting expenses. Updated the filters to use this new structure as well. Added ui feedback for expense submission errors. Next : Once an expense submission is susccessful, the drawer should be closed automatically. Rework the filters sidebar to be more user friendly. Rework the details card : make a better details expense card accordian for displaying all data, and the option to update an expense.