created the stats frontend, and Stats routes and Service
Created the Stats page with implementation for both backend and frontend, with sync logic. Created the monthlySpendingStats component with carousels and data visualisation per category and per user. Added the ability to switch months :groups of 3. Next : Refine the component for visualising data from the last 5 months :setup backend logic for this, add the ability to refine the selection with filters for type of display, category toggles, spender toggles, etc.