settings front and back emplementation + integration with the app + reworking the stats filters
Implemented the settings page, with all the options that the user update as they see fit them best. This includes preferences like currency, formats, visualisation color and type, among others. The integration with the app is done. I also reworked the filters drawer in the stats page, works seemlessly now, as i initially envisionned it. Next : Implement the desktop app layout, then it's time for optimisations and SEO.